Creates framework for different pages and makes cards clickable
