+++ app/package.json
@@ -6,6 +6,7 @@
     "bootstrap": "^4.5.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "0.9.5"
   },
   "devDependencies": {},
