+++ app/src/App.js
@@ -1,36 +1,38 @@
 import React, { Component } from 'react';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+} from "react-router-dom";
 import logo from './logo.svg';
 import './App.css';
-import Widget from './Widget';
+
+import Dashboard from './pages/Dashboard';
+import TodoList from './pages/TodoList';
 
 class App extends Component {
   render() {
-    // Create widgets
-    const widgets = [];
-    for (let i = 0; i < 6; i++) {
-      widgets.push(<Widget name="Small Widget" size="small" />)
-    }
-    for (let i = 0; i < 4; i++) {
-      widgets.push(<Widget name="Medium Widget" size="medium" />)
-    }
-    for (let i = 0; i < 2; i++) {
-      widgets.push(<Widget name="Large Widget" size="large" />)
-    }
-
     return (
-      <div className="App">
-        <div className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h2>Day In Review</h2>
-        </div>
-        {/* Main Body */}
-        <div className="container-fluid App-body">
-          {/* Top Row */}
-          <div className="row">
-            {widgets}
+      <Router>
+        <div className="App">
+          <div className="App-header">
+            <img src={logo} className="App-logo" alt="logo" />
+            <h2>Day In Review</h2>
+          </div>
+          {/* Main Body */}
+          <div className="container-fluid App-body">
+            {/* Pages */}
+            <Switch>
+              <Route exact path="/">
+                <Dashboard />
+              </Route>
+              <Route exact path="/todo">
+                <TodoList />
+              </Route>
+            </Switch>
           </div>
         </div>
-      </div>
+      </Router>
     );
   }
 }
