+++ app/src/pages/TodoList.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import './TodoList.css';
+
+class TodoList extends Component {
+  render() {
+    return(
+      <h2>Todo List Page</h2>
+    );
+  }
+}
+
+export default TodoList;
